requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pymongo==4.11.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.41.3
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1

numpy~=2.2.4
requests~=2.32.3
spacy~=3.8.4
pytest~=8.3.5
networkx~=3.4.2
community~=1.0.0b1
bs4~=0.0.2
beautifulsoup4~=4.13.3